Clarify some language and link to reference material
This clarifies some of the language at the start of the document to make sure that this
doesn't come across as a laundry list of things that have to get done, and instead
encourages everyone to be involved/engaged with the different stages that work goes
through. I also added some additional links to reference material for further reading.